refactor: 정적 데이터 상수로 분리
