+++ apps/landing/src/widgets/landing/constants/hero.ts
@@ -1,0 +1,6 @@
+export const PAIN_POINTS = [
+  { emoji: '😮‍💨', translationKey: 'hero.painPoints.visa' },
+  { emoji: '🤷', translationKey: 'hero.painPoints.e7' },
+  { emoji: '😩', translationKey: 'hero.painPoints.job' },
+  { emoji: '😶‍🌫️', translationKey: 'hero.painPoints.deadline' },
+] as const;
