+++ apps/landing/src/widgets/landing/ui/todo-management-section/todo-management-preview.tsx
@@ -1,37 +1,8 @@
 import { TodoCheckIcon, TodoIcon } from '@kds/icons';
 
-import * as styles from './todo-management-preview.css';
+import { TODO_ITEMS } from '@widgets/landing/constants';
 
-const TODO_ITEMS = [
-  {
-    id: 'opt-application',
-    label: 'Submit OPT Application',
-    due: 'D-1',
-    completed: false,
-    urgent: true,
-  },
-  {
-    id: 'arc-documents',
-    label: 'Prepare ARC Documents',
-    due: 'D-5',
-    completed: false,
-    urgent: false,
-  },
-  {
-    id: 'health-insurance',
-    label: 'Get Health Insurance',
-    due: 'D-10',
-    completed: false,
-    urgent: false,
-  },
-  {
-    id: 'university-registration',
-    label: 'Register at University',
-    due: 'Done',
-    completed: true,
-    urgent: false,
-  },
-] as const;
+import * as styles from './todo-management-preview.css';
 
 const TodoManagementPreview = () => {
   return (
